backend/python/requirements.txt
fastapi
uvicorn
torch
torchvision
transformers
pgvector
psycopg2-binary
Pillow